requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
pinecone-client[grpc]>=3.0.0
langchain_community
langchain_openai
langchain_experimental
langchain_core
setuptools
gunicorn